Improves formatting and adds failure checks

+++ FullReinstall.sh
@@ -17,14 +17,15 @@ if pgrep -x "DiscordCanary" > /dev/null; then
     read -p "
 Warning - Unable to proceed!
 Discord Canary is still running, and must be closed to reinstall.
-Press Enter to quit Discord. "
+Press Enter to quit Discord.
+"
 	killall DiscordCanary
 fi
 
 
 # Remove the directories if they already exist
 if [ -d "powercord" ]; then
-	echo "Removing DiscordCanary/"
+	echo "Removing powercord/"
 	rm -rf powercord
 fi
 if [ -d "DiscordCanary" ]; then
@@ -34,42 +35,62 @@ fi
 # Sometimes, there are a few files that are stubborn to remove.
 if [ -d "DiscordCanary" ]; then
 	echo ">:( Forcibly removing DiscordCanary/"
-	sudo rm -rf DiscordCanary
+	sudo rm -rfv DiscordCanary # You may need to enter a password here.
 fi
 
-
 # Download and unpack Canary
-echo "Downloading Discord Canary..."
+echo "\nDownloading Discord Canary..."
 wget --show-progress --quiet "https://discordapp.com/api/download/canary?platform=linux&format=tar.gz" -O discord.tar.gz
-echo "Unpacking tar.gz..."
-tar -xf discord.tar.gz
-rm discord.tar.gz
+# Did that work?
+if [ $? -eq 0 ]; then
+	echo "Unpacking..."
+	tar -xf discord.tar.gz
+	rm discord.tar.gz
+else
+	echo "Unable to download, skipping.\n"
+	skip_plug="true"
+fi
 
 # Clone the powercord repository
-echo "Cloning powercord..."
+echo "\nCloning powercord..."
 git clone https://github.com/powercord-org/powercord 1>/dev/null
+# Did that fail?
+if [ $? -ne 0 ]; then skip_plug="true"; fi
 
-# We have to patch the powercord injector plug in order to recognise our custom install folder automatically.
-# If this is not done, the user is prompted to manually specify the DiscordCanary folder we just downloaded.
-INSTALL_DIR="$(pwd)/DiscordCanary"
+# If either Discord or Powercord did not download, do not attempt to plug.
+if [ "$skip_plug" != "true" ];  then
+	# All good, proceed!
+
+	# We have to patch the powercord injector plug in order to recognise our custom install folder automatically.
+	# If this is not done, the user is prompted to manually specify the DiscordCanary folder we just downloaded.
+	INSTALL_DIR="$(pwd)/DiscordCanary"
+	echo "
+	Patching install script to recognise our folderpath automatically.
+	If this fails, it will ask you for the location of Canary.
+	In the event that you get prompted, paste in:
+	\"$INSTALL_DIR\"
+	"
+	sed --in-place "/^const KnownLinuxPaths.*/a\ \ '$INSTALL_DIR'," powercord/injectors/linux.js
+
+	# Standard powercord installation procedures.
+	echo "Installing powercord"
+	cd powercord
+	npm i
+	npm run unplug # Just in case, run unplug first.
+	npm run plug
+	cd ..
+else
+	echo "\nSkipping plug of Powercord due to one or more failed steps.
+Note: You will have to plug Powercord manually after correcting the issue(s).
+Please refer to https://powercord.dev/installation for how to do this.\n"
+fi
+
 echo "
-Patching install script to recognise our folderpath automatically.
-If this fails, it will ask you for the location of Canary.
-In the event that you get prompted, paste in: \"$INSTALL_DIR\"
+All done! :)
 "
-sed --in-place "/^const KnownLinuxPaths.*/a\ \ '$INSTALL_DIR'," powercord/injectors/linux.js
 
-# Standard powercord installation procedures.
-echo "Installing powercord"
-cd powercord
-npm i
-npm run unplug # Just in case, run unplug first.
-npm run plug
-cd ..
-
-echo "
-All done!
-"
+# Autorun the Theme Manager
+./ThemeManager.sh
 
 # Start Discord automatically :)
 ./DiscordCanary/DiscordCanary </dev/null &>/dev/null &
